Refactor OpenWeatherClient
The HDF5 cannot handle the geometry object from the original GeoDataFrame,
so the coordinates are kept in the lat and lon columns, since that is all that is needed.

To store the data according to the data model a split_on column is needd to create
a DataFrame for each city, from which the weather is retrieved.