Add logging to all client classes
Recording of metadata and errors added to all client classes.
datastore class was also refactored to handle the storage of metadata and errors according the data model.

To handle the extra information, the ClientResult class was implemented.
This class abstracts the results, metadata and errors produced by the clients
and delivers them to the data store.